Stage1: add a weight annotation to arrows in surface language
- No parsing rule
- No type checking

Open: the annotation is in type `HsType` but not in type
`Type` (=`TcType`). This is because `Type` is used to represent the
types of Core. Depending on how the type-checker works, I may have to
propagate the weight annotation in `Type` as well. In this case, it will
appear in Core, but Core will simply ignore it (for the moment:
eventually Core will understand linear types).